+++ src/sll/api/serial_api.c
@@ -14,14 +14,10 @@
 
 
 
-#define ENCODE_SIGNED_INTEGER(f,n) _encode_integer((f),((n)<0?((~(n))<<1)|1:(n)<<1))
-
-
-
-static sll_size_t _decode_integer(sll_file_t* rf,sll_bool_t* e){
-	sll_read_char_t c=sll_file_read_char(rf);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_size_t sll_decode_integer(sll_file_t* f,sll_bool_t* err){
+	sll_read_char_t c=sll_file_read_char(f);
 	if (c==SLL_END_OF_DATA){
-		*e=1;
+		*err=1;
 		return 0;
 	}
 	sll_size_t v=0;
@@ -29,9 +25,9 @@ static sll_size_t _decode_integer(sll_file_t* rf,sll_bool_t* e){
 	while ((c&0x80)&&s<56){
 		v|=((sll_size_t)(c&0x7f))<<s;
 		s+=7;
-		c=sll_file_read_char(rf);
+		c=sll_file_read_char(f);
 		if (c==SLL_END_OF_DATA){
-			*e=1;
+			*err=1;
 			return 0;
 		}
 	}
@@ -40,23 +36,9 @@ static sll_size_t _decode_integer(sll_file_t* rf,sll_bool_t* e){
 
 
 
-static sll_integer_t _decode_signed_integer(sll_file_t* rf,sll_bool_t* e){
-	sll_size_t v=_decode_integer(rf,e);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_integer_t sll_decode_signed_integer(sll_file_t* f,sll_bool_t* err){
+	sll_size_t v=sll_decode_integer(f,err);
 	return (v>>1)^(-((sll_integer_t)(v&1)));
-}
-
-
-
-static void _encode_integer(sll_file_t* f,sll_size_t v){
-	sll_char_t bf[9];
-	sll_string_length_t i=0;
-	while (i<8&&v>0x7f){
-		bf[i]=(v&0x7f)|0x80;
-		v>>=7;
-		i++;
-	}
-	bf[i]=(sll_char_t)v;
-	sll_file_write(f,bf,i+1);
 }
 
 
@@ -66,14 +48,14 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_decode_object(sll_file_t* f)
 		return SLL_ACQUIRE_STATIC_INT(0);
 	}
 	sll_bool_t err=0;
-	sll_object_type_t t=(sll_object_type_t)_decode_integer(f,&err);
+	sll_object_type_t t=(sll_object_type_t)sll_decode_integer(f,&err);
 	if (err){
 		return SLL_ACQUIRE_STATIC_INT(0);
 	}
 	switch (t){
 		case SLL_OBJECT_TYPE_INT:
 			{
-				sll_integer_t v=_decode_signed_integer(f,&err);
+				sll_integer_t v=sll_decode_signed_integer(f,&err);
 				return (err?SLL_ACQUIRE_STATIC_INT(0):SLL_FROM_INT(v));
 			}
 		case SLL_OBJECT_TYPE_FLOAT:
@@ -105,7 +87,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_decode_object(sll_file_t* f)
 			{
 				sll_object_t* o=SLL_CREATE();
 				o->t=SLL_OBJECT_TYPE_ARRAY;
-				sll_array_length_t l=(sll_array_length_t)_decode_integer(f,&err);
+				sll_array_length_t l=(sll_array_length_t)sll_decode_integer(f,&err);
 				if (err){
 					SLL_INIT_ARRAY(&(o->dt.a));
 				}
@@ -123,7 +105,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_decode_object(sll_file_t* f)
 			{
 				sll_object_t* o=SLL_CREATE();
 				o->t=SLL_OBJECT_TYPE_MAP;
-				sll_map_length_t l=(sll_map_length_t)_decode_integer(f,&err);
+				sll_map_length_t l=(sll_map_length_t)sll_decode_integer(f,&err);
 				if (err){
 					SLL_INIT_MAP(&(o->dt.m));
 				}
@@ -148,7 +130,7 @@ __SLL_EXTERNAL sll_bool_t sll_decode_string(sll_file_t* f,sll_string_t* o){
 		goto _error;
 	}
 	sll_bool_t err=0;
-	o->l=(sll_string_length_t)_decode_integer(f,&err);
+	o->l=(sll_string_length_t)sll_decode_integer(f,&err);
 	if (err){
 		goto _error;
 	}
@@ -220,6 +202,26 @@ _error:
 
 
 
+__SLL_EXTERNAL void sll_encode_integer(sll_file_t* f,sll_size_t v){
+	sll_char_t bf[9];
+	sll_string_length_t i=0;
+	while (i<8&&v>0x7f){
+		bf[i]=(v&0x7f)|0x80;
+		v>>=7;
+		i++;
+	}
+	bf[i]=(sll_char_t)v;
+	sll_file_write(f,bf,i+1);
+}
+
+
+
+__SLL_EXTERNAL void sll_encode_signed_integer(sll_file_t* f,sll_integer_t v){
+	sll_encode_integer(f,(v<0?((~v)<<1)|1:v<<1));
+}
+
+
+
 __SLL_EXTERNAL void sll_encode_object(sll_file_t* f,sll_object_t*const* a,sll_arg_count_t ac){
 	if (!f){
 		return;
@@ -228,10 +230,10 @@ __SLL_EXTERNAL void sll_encode_object(sll_file_t* f,sll_object_t*const* a,sll_ar
 		sll_object_t* k=*a;
 		a++;
 		ac--;
-		_encode_integer(f,SLL_OBJECT_GET_TYPE(k));
+		sll_encode_integer(f,SLL_OBJECT_GET_TYPE(k));
 		switch (SLL_OBJECT_GET_TYPE(k)){
 			case SLL_OBJECT_TYPE_INT:
-				ENCODE_SIGNED_INTEGER(f,k->dt.i);
+				sll_encode_signed_integer(f,k->dt.i);
 				break;
 			case SLL_OBJECT_TYPE_FLOAT:
 				sll_file_write(f,&(k->dt.f),sizeof(sll_float_t));
@@ -245,11 +247,11 @@ __SLL_EXTERNAL void sll_encode_object(sll_file_t* f,sll_object_t*const* a,sll_ar
 			case SLL_OBJECT_TYPE_ARRAY:
 			case SLL_OBJECT_TYPE_MAP_KEYS:
 			case SLL_OBJECT_TYPE_MAP_VALUES:
-				_encode_integer(f,k->dt.a.l);
+				sll_encode_integer(f,k->dt.a.l);
 				sll_encode_object(f,k->dt.a.v,k->dt.a.l);
 				break;
 			case SLL_OBJECT_TYPE_MAP:
-				_encode_integer(f,k->dt.m.l);
+				sll_encode_integer(f,k->dt.m.l);
 				sll_encode_object(f,k->dt.m.v,k->dt.m.l<<1);
 				break;
 			default:
@@ -264,7 +266,7 @@ __SLL_EXTERNAL void sll_encode_string(sll_file_t* f,const sll_string_t* s){
 	if (!f){
 		return;
 	}
-	_encode_integer(f,s->l);
+	sll_encode_integer(f,s->l);
 	if (s->l<STRING_COMPRESSION_MIN_LENGTH){
 		sll_file_write(f,s->v,s->l*sizeof(sll_char_t));
 		return;
@@ -344,6 +346,33 @@ __SLL_EXTERNAL void sll_encode_string(sll_file_t* f,const sll_string_t* s){
 
 
 
+__API_FUNC(serial_decode_float){
+	sll_file_t* f=sll_file_from_handle(a);
+	if (!f){
+		return 0;
+	}
+	sll_float_t v;
+	return (sll_file_read(f,&v,sizeof(sll_float_t))==sizeof(sll_float_t)?v:0);
+}
+
+
+
+__API_FUNC(serial_decode_integer){
+	sll_bool_t err=0;
+	sll_size_t o=sll_decode_integer(sll_file_from_handle(a),&err);
+	return (err?0:(sll_integer_t)o);
+}
+
+
+
+__API_FUNC(serial_decode_signed_integer){
+	sll_bool_t err=0;
+	sll_integer_t o=sll_decode_signed_integer(sll_file_from_handle(a),&err);
+	return (err?0:o);
+}
+
+
+
 __API_FUNC(serial_decode_object){
 	return sll_decode_object(sll_file_from_handle(a));
 }
@@ -352,6 +381,24 @@ __API_FUNC(serial_decode_object){
 
 __API_FUNC(serial_decode_string){
 	sll_decode_string(sll_file_from_handle(a),out);
+}
+
+
+
+__API_FUNC(serial_encode_float){
+	sll_file_write(sll_file_from_handle(a),&b,sizeof(sll_float_t));
+}
+
+
+
+__API_FUNC(serial_encode_integer){
+	sll_encode_integer(sll_file_from_handle(a),(sll_size_t)b);
+}
+
+
+
+__API_FUNC(serial_encode_signed_integer){
+	sll_encode_signed_integer(sll_file_from_handle(a),b);
 }