+++ src/sll/include/sll/api/serial.h
@@ -14,6 +14,32 @@
 
 /**
  * \flags check_output func
+ * \name sll_decode_integer
+ * \group serial-api
+ * \desc Docs!
+ * \arg sll_file_t* f
+ * \arg sll_bool_t* err
+ * \ret sll_size_t
+ */
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_size_t sll_decode_integer(sll_file_t* f,sll_bool_t* err);
+
+
+
+/**
+ * \flags check_output func
+ * \name sll_decode_signed_integer
+ * \group serial-api
+ * \desc Docs!
+ * \arg sll_file_t* f
+ * \arg sll_bool_t* err
+ * \ret sll_integer_t
+ */
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_integer_t sll_decode_signed_integer(sll_file_t* f,sll_bool_t* err);
+
+
+
+/**
+ * \flags check_output func
  * \name sll_decode_object
  * \group serial-api
  * \desc Docs!
@@ -38,7 +64,31 @@ __SLL_EXTERNAL sll_bool_t sll_decode_string(sll_file_t* f,sll_string_t* o);
 
 
 /**
- * \flags check_output func
+ * \flags func
+ * \name sll_encode_integer
+ * \group serial-api
+ * \desc Docs!
+ * \arg sll_file_t* f
+ * \arg sll_size_t v
+ */
+__SLL_EXTERNAL void sll_encode_integer(sll_file_t* f,sll_size_t v);
+
+
+
+/**
+ * \flags func
+ * \name sll_encode_signed_integer
+ * \group serial-api
+ * \desc Docs!
+ * \arg sll_file_t* f
+ * \arg sll_integer_t v
+ */
+__SLL_EXTERNAL void sll_encode_signed_integer(sll_file_t* f,sll_integer_t v);
+
+
+
+/**
+ * \flags func
  * \name sll_encode_object
  * \group serial-api
  * \desc Docs!
@@ -64,6 +114,42 @@ __SLL_EXTERNAL void sll_encode_string(sll_file_t* f,const sll_string_t* s);
 
 /**
  * \flags api func
+ * \name sll_api_serial_decode_float
+ * \group serial-api
+ * \desc Docs!
+ * \arg I
+ * \ret F
+ */
+__API_FUNC_DECL(serial_decode_float);
+
+
+
+/**
+ * \flags api func
+ * \name sll_api_serial_decode_integer
+ * \group serial-api
+ * \desc Docs!
+ * \arg I
+ * \ret I
+ */
+__API_FUNC_DECL(serial_decode_integer);
+
+
+
+/**
+ * \flags api func
+ * \name sll_api_serial_decode_signed_integer
+ * \group serial-api
+ * \desc Docs!
+ * \arg I
+ * \ret I
+ */
+__API_FUNC_DECL(serial_decode_signed_integer);
+
+
+
+/**
+ * \flags api func
  * \name sll_api_serial_decode_object
  * \group serial-api
  * \desc Docs!
@@ -83,6 +169,42 @@ __API_FUNC_DECL(serial_decode_object);
  * \ret S
  */
 __API_FUNC_DECL(serial_decode_string);
+
+
+
+/**
+ * \flags api func
+ * \name sll_api_serial_encode_float
+ * \group serial-api
+ * \desc Docs!
+ * \arg I
+ * \arg F
+ */
+__API_FUNC_DECL(serial_encode_float);
+
+
+
+/**
+ * \flags api func
+ * \name sll_api_serial_encode_integer
+ * \group serial-api
+ * \desc Docs!
+ * \arg I
+ * \arg I
+ */
+__API_FUNC_DECL(serial_encode_integer);
+
+
+
+/**
+ * \flags api func
+ * \name sll_api_serial_encode_signed_integer
+ * \group serial-api
+ * \desc Docs!
+ * \arg I
+ * \arg I
+ */
+__API_FUNC_DECL(serial_encode_signed_integer);