+++ src/sll/io/write.c
@@ -12,19 +12,6 @@
 
 
 #define WRITE_FIELD(f,wf) sll_file_write((wf),&(f),sizeof((f)))
-#define WRITE_SIGNED_INTEGER(wf,n) _write_integer((wf),((n)<0?((~(n))<<1)|1:(n)<<1))
-
-
-
-static void _write_integer(sll_file_t* wf,sll_size_t v){
-	unsigned int i=0;
-	while (i<8&&v>0x7f){
-		sll_file_write_char(wf,(v&0x7f)|0x80);
-		v>>=7;
-		i++;
-	}
-	sll_file_write_char(wf,(sll_char_t)v);
-}
 
 
 
@@ -36,7 +23,7 @@ static const sll_node_t* _write_node(sll_file_t* wf,const sll_node_t* o){
 		}
 		WRITE_FIELD(o->t,wf);
 		if (o->t==SLL_NODE_TYPE_DBG){
-			_write_integer(wf,o->dt.s+1);
+			sll_encode_integer(wf,o->dt.s+1);
 		}
 		o++;
 	}
@@ -46,18 +33,18 @@ static const sll_node_t* _write_node(sll_file_t* wf,const sll_node_t* o){
 			WRITE_FIELD(o->dt.c,wf);
 			return o+1;
 		case SLL_NODE_TYPE_INT:
-			WRITE_SIGNED_INTEGER(wf,o->dt.i);
+			sll_encode_signed_integer(wf,o->dt.i);
 			return o+1;
 		case SLL_NODE_TYPE_FLOAT:
 			WRITE_FIELD(o->dt.f,wf);
 			return o+1;
 		case SLL_NODE_TYPE_STRING:
 		case SLL_NODE_TYPE_FIELD:
-			_write_integer(wf,o->dt.s);
+			sll_encode_integer(wf,o->dt.s);
 			return o+1;
 		case SLL_NODE_TYPE_ARRAY:
 			{
-				_write_integer(wf,o->dt.al);
+				sll_encode_integer(wf,o->dt.al);
 				sll_array_length_t l=o->dt.al;
 				o++;
 				while (l){
@@ -68,7 +55,7 @@ static const sll_node_t* _write_node(sll_file_t* wf,const sll_node_t* o){
 			}
 		case SLL_NODE_TYPE_MAP:
 			{
-				_write_integer(wf,o->dt.ml);
+				sll_encode_integer(wf,o->dt.ml);
 				sll_map_length_t l=o->dt.ml;
 				o++;
 				while (l){
@@ -78,18 +65,18 @@ static const sll_node_t* _write_node(sll_file_t* wf,const sll_node_t* o){
 				return o;
 			}
 		case SLL_NODE_TYPE_IDENTIFIER:
-			_write_integer(wf,o->dt.id);
+			sll_encode_integer(wf,o->dt.id);
 			return o+1;
 		case SLL_NODE_TYPE_FUNCTION_ID:
-			_write_integer(wf,o->dt.fn_id);
+			sll_encode_integer(wf,o->dt.fn_id);
 			return o+1;
 		case SLL_NODE_TYPE_FUNC:
 		case SLL_NODE_TYPE_INTERNAL_FUNC:
 			{
-				_write_integer(wf,o->dt.fn.ac);
-				_write_integer(wf,o->dt.fn.id);
+				sll_encode_integer(wf,o->dt.fn.ac);
+				sll_encode_integer(wf,o->dt.fn.id);
 				if (o->t==SLL_NODE_TYPE_FUNC){
-					_write_integer(wf,o->dt.fn.sc);
+					sll_encode_integer(wf,o->dt.fn.sc);
 				}
 				sll_arg_count_t l=o->dt.fn.ac;
 				o++;
@@ -107,8 +94,8 @@ static const sll_node_t* _write_node(sll_file_t* wf,const sll_node_t* o){
 		case SLL_NODE_TYPE_FOR_MAP:
 		case SLL_NODE_TYPE_WHILE_MAP:
 			{
-				_write_integer(wf,o->dt.l.ac);
-				_write_integer(wf,o->dt.l.sc);
+				sll_encode_integer(wf,o->dt.l.ac);
+				sll_encode_integer(wf,o->dt.l.sc);
 				sll_arg_count_t l=o->dt.l.ac;
 				o++;
 				while (l){
@@ -119,8 +106,8 @@ static const sll_node_t* _write_node(sll_file_t* wf,const sll_node_t* o){
 			}
 		case SLL_NODE_TYPE_DECL:
 			{
-				_write_integer(wf,o->dt.d.ac);
-				_write_integer(wf,o->dt.d.nm+1);
+				sll_encode_integer(wf,o->dt.d.ac);
+				sll_encode_integer(wf,o->dt.d.nm+1);
 				sll_arg_count_t l=o->dt.d.ac;
 				o++;
 				while (l){
@@ -130,7 +117,7 @@ static const sll_node_t* _write_node(sll_file_t* wf,const sll_node_t* o){
 				return o;
 			}
 	}
-	_write_integer(wf,o->dt.ac);
+	sll_encode_integer(wf,o->dt.ac);
 	sll_arg_count_t l=o->dt.ac;
 	o++;
 	while (l){
@@ -147,23 +134,23 @@ __SLL_EXTERNAL void sll_write_assembly(sll_file_t* wf,const sll_assembly_data_t*
 	sll_file_write(wf,&n,sizeof(magic_number_t));
 	sll_version_t v=SLL_VERSION;
 	sll_file_write(wf,&v,sizeof(sll_version_t));
-	_write_integer(wf,a_dt->tm);
-	_write_integer(wf,a_dt->ic);
-	_write_integer(wf,a_dt->vc);
-	_write_integer(wf,a_dt->ft.l);
+	sll_encode_integer(wf,a_dt->tm);
+	sll_encode_integer(wf,a_dt->ic);
+	sll_encode_integer(wf,a_dt->vc);
+	sll_encode_integer(wf,a_dt->ft.l);
 	for (sll_function_index_t i=0;i<a_dt->ft.l;i++){
-		_write_integer(wf,(a_dt->ft.dt+i)->i);
-		_write_integer(wf,(a_dt->ft.dt+i)->ac);
-		_write_integer(wf,(a_dt->ft.dt+i)->nm);
+		sll_encode_integer(wf,(a_dt->ft.dt+i)->i);
+		sll_encode_integer(wf,(a_dt->ft.dt+i)->ac);
+		sll_encode_integer(wf,(a_dt->ft.dt+i)->nm);
 	}
-	_write_integer(wf,a_dt->st.l);
+	sll_encode_integer(wf,a_dt->st.l);
 	for (sll_string_index_t i=0;i<a_dt->st.l;i++){
 		sll_encode_string(wf,a_dt->st.dt+i);
 	}
-	_write_integer(wf,a_dt->dbg.l);
+	sll_encode_integer(wf,a_dt->dbg.l);
 	for (sll_instruction_index_t i=0;i<a_dt->dbg.l;i++){
-		_write_integer(wf,(a_dt->dbg.dt+i)->ii);
-		_write_integer(wf,ROTATE_BITS64((a_dt->dbg.dt+i)->ln,1));
+		sll_encode_integer(wf,(a_dt->dbg.dt+i)->ii);
+		sll_encode_integer(wf,ROTATE_BITS64((a_dt->dbg.dt+i)->ln,1));
 	}
 	const sll_assembly_instruction_t* ai=a_dt->h;
 	for (sll_instruction_index_t i=0;i<a_dt->ic;i++){
@@ -173,7 +160,7 @@ __SLL_EXTERNAL void sll_write_assembly(sll_file_t* wf,const sll_assembly_data_t*
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_CALL_ZERO:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_CALL_ONE:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_RET_INT:
-				WRITE_SIGNED_INTEGER(wf,ai->dt.i);
+				sll_encode_signed_integer(wf,ai->dt.i);
 				break;
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_PUSH_FLOAT:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_RET_FLOAT:
@@ -199,10 +186,10 @@ __SLL_EXTERNAL void sll_write_assembly(sll_file_t* wf,const sll_assembly_data_t*
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_JI:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_JNI:
 				if (SLL_ASSEMBLY_INSTRUCTION_FLAG_IS_RELATIVE(ai)){
-					WRITE_SIGNED_INTEGER(wf,ai->dt.rj);
+					sll_encode_signed_integer(wf,ai->dt.rj);
 				}
 				else{
-					_write_integer(wf,ai->dt.rj);
+					sll_encode_integer(wf,ai->dt.rj);
 				}
 				break;
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_LOAD:
@@ -218,20 +205,20 @@ __SLL_EXTERNAL void sll_write_assembly(sll_file_t* wf,const sll_assembly_data_t*
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_RET_VAR:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_DEL:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_LOAD_DEL:
-				_write_integer(wf,ai->dt.v);
+				sll_encode_integer(wf,ai->dt.v);
 				break;
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_LOADS:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_LOOKUP_STR:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_PRINT_STR:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_RET_STR:
-				_write_integer(wf,ai->dt.s);
+				sll_encode_integer(wf,ai->dt.s);
 				break;
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_PACK:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_JT:
-				_write_integer(wf,ai->dt.al);
+				sll_encode_integer(wf,ai->dt.al);
 				break;
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_MAP:
-				_write_integer(wf,ai->dt.ml);
+				sll_encode_integer(wf,ai->dt.ml);
 				break;
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_NOT:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_INC:
@@ -250,13 +237,13 @@ __SLL_EXTERNAL void sll_write_assembly(sll_file_t* wf,const sll_assembly_data_t*
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_ASSIGN_TWO:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_ASSIGN_THREE:
 				if (SLL_ASSEMBLY_INSTRUCTION_FLAG_IS_INPLACE(ai)){
-					_write_integer(wf,ai->dt.v);
+					sll_encode_integer(wf,ai->dt.v);
 				}
 				break;
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_ACCESS_VAR:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_ASSIGN_VAR_ACCESS:
-				_write_integer(wf,ai->dt.va.v);
-				_write_integer(wf,ai->dt.va.l);
+				sll_encode_integer(wf,ai->dt.va.v);
+				sll_encode_integer(wf,ai->dt.va.l);
 				break;
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_CAST_TYPE:
 				sll_file_write_char(wf,ai->dt.t);
@@ -265,7 +252,7 @@ __SLL_EXTERNAL void sll_write_assembly(sll_file_t* wf,const sll_assembly_data_t*
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_CALL_POP:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_DECL:
 			case SLL_ASSEMBLY_INSTRUCTION_TYPE_NEW:
-				_write_integer(wf,ai->dt.ac);
+				sll_encode_integer(wf,ai->dt.ac);
 				break;
 		}
 		ai++;
@@ -288,42 +275,42 @@ __SLL_EXTERNAL void sll_write_compiled_node(sll_file_t* wf,const sll_compilation
 	sll_file_write(wf,&n,sizeof(magic_number_t));
 	sll_version_t v=SLL_VERSION;
 	sll_file_write(wf,&v,sizeof(sll_version_t));
-	_write_integer(wf,c_dt->tm);
+	sll_encode_integer(wf,c_dt->tm);
 	for (sll_identifier_index_t i=0;i<SLL_MAX_SHORT_IDENTIFIER_LENGTH;i++){
 		const sll_identifier_list_t* l=c_dt->idt.s+i;
-		_write_integer(wf,l->l);
+		sll_encode_integer(wf,l->l);
 		for (sll_identifier_list_length_t j=0;j<l->l;j++){
-			_write_integer(wf,(l->dt+j)->sc);
-			_write_integer(wf,(l->dt+j)->i);
+			sll_encode_integer(wf,(l->dt+j)->sc);
+			sll_encode_integer(wf,(l->dt+j)->i);
 		}
 	}
-	_write_integer(wf,c_dt->idt.ill);
+	sll_encode_integer(wf,c_dt->idt.ill);
 	for (sll_identifier_list_length_t i=0;i<c_dt->idt.ill;i++){
-		_write_integer(wf,(c_dt->idt.il+i)->sc);
-		_write_integer(wf,(c_dt->idt.il+i)->i);
+		sll_encode_integer(wf,(c_dt->idt.il+i)->sc);
+		sll_encode_integer(wf,(c_dt->idt.il+i)->i);
 	}
-	_write_integer(wf,c_dt->et.l);
+	sll_encode_integer(wf,c_dt->et.l);
 	for (sll_export_table_length_t i=0;i<c_dt->et.l;i++){
-		_write_integer(wf,*(c_dt->et.dt+i));
+		sll_encode_integer(wf,*(c_dt->et.dt+i));
 	}
-	_write_integer(wf,c_dt->ft.l);
+	sll_encode_integer(wf,c_dt->ft.l);
 	for (sll_function_index_t i=0;i<c_dt->ft.l;i++){
 		const sll_function_t* k=*(c_dt->ft.dt+i);
-		_write_integer(wf,k->off);
-		_write_integer(wf,k->al);
-		_write_integer(wf,k->nm+1);
+		sll_encode_integer(wf,k->off);
+		sll_encode_integer(wf,k->al);
+		sll_encode_integer(wf,k->nm+1);
 		for (sll_arg_count_t j=0;j<k->al;j++){
-			_write_integer(wf,k->a[j]);
+			sll_encode_integer(wf,k->a[j]);
 		}
 	}
-	_write_integer(wf,c_dt->st.l);
+	sll_encode_integer(wf,c_dt->st.l);
 	for (sll_string_index_t i=0;i<c_dt->st.l;i++){
 		sll_encode_string(wf,c_dt->st.dt+i);
 	}
-	_write_integer(wf,c_dt->fpt.l);
+	sll_encode_integer(wf,c_dt->fpt.l);
 	for (sll_string_index_t i=0;i<c_dt->fpt.l;i++){
-		_write_integer(wf,*(c_dt->fpt.dt+i));
+		sll_encode_integer(wf,*(c_dt->fpt.dt+i));
 	}
-	_write_integer(wf,c_dt->_n_sc_id);
+	sll_encode_integer(wf,c_dt->_n_sc_id);
 	_write_node(wf,c_dt->h);
 }